+++ src/app/collective-rewards/whitelist/WhitelistGridItem.tsx
@@ -1,5 +1,5 @@
 import { crStatusColorClasses } from '@/app/collective-rewards/BecomeABuilderButton'
-import { AddressOrAlias, reduceAddress } from '@/components/Address'
+import { AddressOrAliasWithCopy } from '@/components/Address'
 import { Badge } from '@/components/Badge'
 import { Popover } from '@/components/Popover'
 import { Paragraph, Span, Typography } from '@/components/Typography'
@@ -7,6 +7,8 @@ import { useRouter } from 'next/navigation'
 import { FC, ReactNode } from 'react'
 import { Jdenticon } from '@/components/Header/Jdenticon'
 import { BuilderProposal } from '@/app/collective-rewards/BuilderContext'
+import { Address } from 'viem'
+import { shortAddress } from '@/lib/utils'
 
 type WhitelistGridItemProps = BuilderProposal
 
@@ -39,14 +41,15 @@ export const WhitelistGridItem: FC<WhitelistGridItemProps> = ({
         <Popover
           content={
             <div className="text-[12px] font-bold mb-1">
-              <p data-testid="builderAddressTooltip">{reduceAddress(address)}</p>
+              <p data-testid="builderAddressTooltip">{shortAddress(address as Address)}</p>
             </div>
           }
           size="small"
           trigger="hover"
+          disabled={!builderName}
         >
           <Typography tagVariant="label" className="font-semibold line-clamp-1 text-wrap">
-            <AddressOrAlias
+            <AddressOrAliasWithCopy
               addressOrAlias={builderName || address}
               clipboard={address}
               className="text-base font-bold"
